backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
faiss-cpu==1.9.0.post1
numpy==1.26.4
pdfplumber==0.11.4
python-pptx==1.0.2
python-dotenv==1.0.1
pydantic==2.9.2
